+++ internal/cli/template/lint.go
@@ -39,6 +39,11 @@ files with the .yaml or .yml extension.`)[1:],
 			if err != nil {
 				return fmt.Errorf("lint paths error: %w", err)
 			}
+			type result struct {
+				target string
+				ok     bool
+			}
+			var lintResults []result
 			var pathLints []pathLint
 			for _, target := range targets {
 				if target == "" {
@@ -47,6 +52,18 @@ files with the .yaml or .yml extension.`)[1:],
 				lints := lintFile(target)
 				if len(lints) > 0 {
 					pathLints = append(pathLints, lints...)
+					lintResults = append(lintResults, result{target, false})
+				} else {
+					lintResults = append(lintResults, result{target, true})
+				}
+			}
+			if c.Bool("verbose") {
+				for _, res := range lintResults {
+					if res.ok {
+						fmt.Fprintf(opts.Stdout, "%v: %v\n", res.target, green("OK"))
+					} else {
+						fmt.Fprintf(opts.Stdout, "%v: %v\n", res.target, red("FAILED"))
+					}
 				}
 			}
 			if len(pathLints) == 0 {
@@ -68,6 +85,7 @@ files with the .yaml or .yml extension.`)[1:],
 var (
 	red    = color.New(color.FgRed).SprintFunc()
 	yellow = color.New(color.FgYellow).SprintFunc()
+	green  = color.New(color.FgGreen).SprintFunc()
 )
 
 type pathLint struct {