+++ internal/bloblang/query/functions_test.go
@@ -3,6 +3,7 @@
 package query
 
 import (
+	"errors"
 	"fmt"
 	"math"
 	"sync"
@@ -492,4 +493,110 @@ func TestRandomIntWithinRange(t *testing.T) {
 	// Create a new random_int function with a max that will overflow
 	_, err = InitFunctionHelper("random_int", tsFn, 0, math.MaxInt64)
 	require.Error(t, err)
+}
+
+func TestErrorFunctions(t *testing.T) {
+	tests := []struct {
+		name           string
+		fn             string
+		err            string
+		componentName  string
+		componentLabel string
+		componentPath  []string
+		exp            string
+	}{
+		{
+			name: "error() returns the message error if set",
+			fn:   "error",
+			err:  "kaboom!",
+			exp:  "kaboom!",
+		},
+		{
+			name: "error() returns null for non-errored messages",
+			fn:   "error",
+		},
+		{
+			name: "errored() returns true for errored messages",
+			fn:   "errored",
+			err:  "kaboom!",
+		},
+		{
+			name: "errored() returns false for non-errored messages",
+			fn:   "errored",
+		},
+		{
+			name:          "error_source_name() returns the component name when the message is errored",
+			fn:            "error_source_name",
+			err:           "kaboom!",
+			componentName: "foobar",
+			exp:           "foobar",
+		},
+		{
+			name: "error_source_name() returns null for non-errored messages",
+			fn:   "error_source_name",
+		},
+		{
+			name:           "error_source_label() returns the component label when the message is errored",
+			fn:             "error_source_label",
+			err:            "kaboom!",
+			componentLabel: "foobar",
+			exp:            "foobar",
+		},
+		{
+			name: "error_source_label() returns null for non-errored messages",
+			fn:   "error_source_label",
+		},
+		{
+			name:          "error_source_path() returns the component path when the message is errored",
+			fn:            "error_source_path",
+			err:           "kaboom!",
+			componentPath: []string{"foo", "bar"},
+			exp:           "foo.bar",
+		},
+		{
+			name: "error_source_path() returns null for non-errored messages",
+			fn:   "error_source_path",
+		},
+	}
+
+	for _, test := range tests {
+		t.Run(test.name, func(t *testing.T) {
+			batch := message.QuickBatch(nil)
+			part := message.NewPart([]byte("foobar"))
+			if test.err != "" {
+				part.ErrorSet(&ComponentError{
+					Err:   errors.New(test.err),
+					Name:  test.componentName,
+					Label: test.componentLabel,
+					Path:  test.componentPath,
+				})
+			}
+			batch = append(batch, part)
+
+			e, err := InitFunctionHelper(test.fn)
+			require.NoError(t, err)
+			res, err := e.Exec(FunctionContext{
+				Index:    0,
+				MsgBatch: batch,
+			})
+			require.NoError(t, err)
+
+			if test.fn == "errored" {
+				b, ok := res.(bool)
+				require.True(t, ok)
+				assert.Equal(t, b, test.err != "")
+				return
+			}
+
+			if test.err == "" {
+				assert.Empty(t, res)
+				return
+			}
+
+			str, ok := res.(string)
+			require.True(t, ok)
+			assert.Equal(t, test.exp, str)
+		})
+	}
+
 }