+++ CHANGELOG.md
@@ -8,7 +8,8 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - Field `label` added to the template tests definitions. (@mihaitodor)
-- Metadata field `label` can now be utilized within a template's `mapping` field to access the label that is associated with the template instantiation in a config.
+- Metadata field `label` can now be utilized within a template's `mapping` field to access the label that is associated with the template instantiation in a config. (@mihaitodor)
+- Flag `--verbose` added to the `benthos lint` and `benthos template lint` commands. (@mihaitodor)
 
 ## 4.42.0 - 2024-11-29