+++ CHANGELOG.md
@@ -3,7 +3,9 @@ Changelog
 
 All notable changes to this project will be documented in this file.
 
-## 4.43.0 - TBD
+## 4.44.0 - TBD
+
+## 4.43.0 - 2025-01-13
 
 ### Added